+++ saleor/discount/tests/test_discounts.py
@@ -6,6 +6,7 @@ import pytest
 from django.utils import timezone
 from prices import Money, TaxedMoney
 
+from ...checkout.fetch import CheckoutLineInfo
 from ...discount.interface import VariantPromotionRuleInfo
 from .. import DiscountValueType, RewardValueType, VoucherType
 from ..models import (
@@ -16,6 +17,7 @@ from ..models import (
     VoucherCustomer,
 )
 from ..utils import (
+    _get_the_cheapest_line,
     activate_voucher_code,
     add_voucher_usage_by_customer,
     deactivate_voucher_code,
@@ -23,6 +25,7 @@ from ..utils import (
     get_discount_name,
     get_discount_translated_name,
     increase_voucher_code_usage_value,
+    is_order_level_voucher,
     remove_voucher_usage_by_customer,
     validate_voucher,
 )
@@ -275,6 +278,15 @@ def test_add_voucher_usage_by_customer_raise_not_applicable(voucher_customer):
     # when & then
     with pytest.raises(NotApplicable):
         add_voucher_usage_by_customer(code, customer_email)
+
+
+def test_add_voucher_usage_by_customer_without_customer_email(voucher):
+    # given
+    code = voucher.codes.first()
+
+    # when & then
+    with pytest.raises(NotApplicable):
+        add_voucher_usage_by_customer(code, None)
 
 
 def test_remove_voucher_usage_by_customer(voucher_customer):
@@ -579,3 +591,82 @@ def test_get_discount_translated_name_no_translations(rule_info):
 
     # then
     assert translated_name is None
+
+
+def test_is_order_level_voucher(voucher):
+    # given
+    voucher.type = VoucherType.ENTIRE_ORDER
+    voucher.save(update_fields=["type"])
+
+    # when
+    result = is_order_level_voucher(voucher)
+
+    # then
+    assert result is True
+
+
+def test_is_order_level_voucher_apply_once_per_order(voucher):
+    # given
+    voucher.type = VoucherType.ENTIRE_ORDER
+    voucher.apply_once_per_order = True
+    voucher.save(update_fields=["type", "apply_once_per_order"])
+
+    # when
+    result = is_order_level_voucher(voucher)
+
+    # then
+    assert result is False
+
+
+def test_is_order_level_voucher_no_voucher(voucher):
+    # when
+    result = is_order_level_voucher(None)
+
+    # then
+    assert result is False
+
+
+@pytest.mark.parametrize(
+    "voucher_type", [VoucherType.SPECIFIC_PRODUCT, VoucherType.SHIPPING]
+)
+def test_is_order_level_voucher_another_type(voucher_type, voucher):
+    # given
+    voucher.type = voucher_type
+    voucher.save(update_fields=["type"])
+
+    # when
+    result = is_order_level_voucher(voucher)
+
+    # then
+    assert result is False
+
+
+def test_get_the_cheapest_line_no_lines_provided():
+    # when
+    line_info = _get_the_cheapest_line(None)
+    # then
+    assert line_info is None
+
+
+def test_get_the_cheapest_line(checkout_with_items, channel_USD):
+    # given
+    lines = [
+        CheckoutLineInfo(
+            line=line,
+            channel_listing=line.variant.channel_listings.first(),
+            collections=[],
+            product=line.variant.product,
+            variant=line.variant,
+            discounts=list(line.discounts.all()),
+            rules_info=[],
+            product_type=line.variant.product.product_type,
+            channel=channel_USD,
+            voucher=None,
+            voucher_code=None,
+        )
+        for line in checkout_with_items.lines.all()
+    ]
+    # when
+    line_info = _get_the_cheapest_line(lines)
+    # then
+    assert line_info == lines[0]