+++ saleor/graphql/checkout/dataloaders.py
@@ -7,7 +7,6 @@ from promise import Promise
 from ...checkout.fetch import (
     CheckoutInfo,
     CheckoutLineInfo,
-    apply_voucher_to_checkout_line,
 )
 from ...checkout.models import Checkout, CheckoutLine, CheckoutMetadata
 from ...checkout.problems import (
@@ -23,6 +22,7 @@ from ...checkout.problems import (
 from ...core.db.connection import allow_writer_in_context
 from ...discount import VoucherType
 from ...discount.interface import VariantPromotionRuleInfo
+from ...discount.utils import apply_voucher_to_line
 from ...payment.models import TransactionItem
 from ...product.models import ProductChannelListing
 from ...warehouse.models import Stock
@@ -143,6 +143,8 @@ class CheckoutLinesInfoByCheckoutTokenLoader(DataLoader[str, list[CheckoutLineIn
                                 tax_class=tax_class_map[line.variant_id],
                                 channel=channels[checkout.channel_id],
                                 rules_info=rules_info_map[line.id],
+                                voucher=None,
+                                voucher_code=None,
                             )
                             for line in lines
                         ]
@@ -159,7 +161,7 @@ class CheckoutLinesInfoByCheckoutTokenLoader(DataLoader[str, list[CheckoutLineIn
                         voucher.type == VoucherType.SPECIFIC_PRODUCT
                         or voucher.apply_once_per_order
                     ):
-                        apply_voucher_to_checkout_line(
+                        apply_voucher_to_line(
                             voucher_info=voucher_info,
                             lines_info=lines_info_map[checkout.pk],
                         )
@@ -637,27 +639,26 @@ class CheckoutMetadataByCheckoutIdLoader(DataLoader[str, CheckoutMetadata]):
         return [checkout_metadata.get(checkout_id) for checkout_id in keys]
 
 
-class ChannelByCheckoutLineIDLoader(DataLoader):
-    context_key = "channel_by_checkout_line"
+class ChannelByCheckoutIDLoader(DataLoader):
+    context_key = "channel_by_checkout"
 
     def batch_load(self, keys):
-        def channel_by_lines(checkout_lines):
-            checkout_ids = [line.checkout_id for line in checkout_lines]
+        def with_checkouts(checkouts):
+            def with_channels(channels):
+                channel_map = {channel.id: channel for channel in channels}
+                return [
+                    channel_map.get(checkout.channel_id) if checkout else None
+                    for checkout in checkouts
+                ]
 
-            def channels_by_checkout(checkouts):
-                channel_ids = [checkout.channel_id for checkout in checkouts]
-
-                return ChannelByIdLoader(self.context).load_many(channel_ids)
-
+            channel_ids = set(checkout.channel_id for checkout in checkouts if checkout)
             return (
-                CheckoutByTokenLoader(self.context)
-                .load_many(checkout_ids)
-                .then(channels_by_checkout)
+                ChannelByIdLoader(self.context)
+                .load_many(channel_ids)
+                .then(with_channels)
             )
 
-        return (
-            CheckoutLineByIdLoader(self.context).load_many(keys).then(channel_by_lines)
-        )
+        return CheckoutByTokenLoader(self.context).load_many(keys).then(with_checkouts)
 
 
 class CheckoutLinesProblemsByCheckoutIdLoader(