+++ saleor/order/utils.py
@@ -15,7 +15,7 @@ from ..core.tracing import traced_atomic_transaction
 from ..core.utils.country import get_active_country
 from ..core.utils.translations import get_translation
 from ..core.weight import zero_weight
-from ..discount import DiscountType
+from ..discount import DiscountType, DiscountValueType
 from ..discount.models import (
     NotApplicable,
     OrderDiscount,
@@ -280,6 +280,7 @@ def create_order_line(
         total_price=total_price,
         undiscounted_total_price=undiscounted_total_price,
         variant=variant,
+        is_price_overridden=price_override is not None,
         **get_tax_class_kwargs_for_order_line(tax_class),
     )
 
@@ -303,6 +304,7 @@ def create_order_line(
         else:
             discount_amount = unit_discount.net
         line.unit_discount = discount_amount
+        line.unit_discount_type = DiscountValueType.FIXED
         line.unit_discount_value = discount_amount.amount
 
         line.save(
@@ -310,6 +312,7 @@ def create_order_line(
                 "unit_discount_amount",
                 "unit_discount_value",
                 "unit_discount_reason",
+                "unit_discount_type",
                 "sale_id",
             ]
         )
@@ -378,6 +381,12 @@ def add_variant_to_order(
         old_quantity = line.quantity
         new_quantity = old_quantity + line_data.quantity
         line_info = OrderLineInfo(line=line, quantity=old_quantity)
+        update_fields: list[str] = []
+        if new_quantity and line_data.price_override is not None:
+            update_line_base_unit_prices_with_custom_price(
+                order, line_data, line, update_fields
+            )
+
         change_order_line_quantity(
             user,
             app,
@@ -387,7 +396,11 @@ def add_variant_to_order(
             channel,
             manager=manager,
             send_event=False,
+            update_fields=update_fields,
         )
+
+        if update_fields:
+            line.save(update_fields=update_fields)
 
         if allocate_stock:
             increase_allocations(
@@ -412,6 +425,38 @@ def add_variant_to_order(
             manager,
             allocate_stock,
         )
+
+
+def update_line_base_unit_prices_with_custom_price(
+    order, line_data, line, update_fields
+):
+    channel = order.channel
+    variant = line_data.variant
+    price_override = line_data.price_override
+    rules_info = line_data.rules_info
+    channel_listing = variant.channel_listings.get(channel=channel)
+
+    line.is_price_overridden = True
+    line.base_unit_price = variant.get_price(
+        channel_listing,
+        price_override=price_override,
+        promotion_rules=(
+            [rule_info.rule for rule_info in rules_info] if rules_info else None
+        ),
+    )
+    line.undiscounted_base_unit_price_amount = price_override
+    line.undiscounted_unit_price_gross_amount = price_override
+    line.undiscounted_unit_price_net_amount = price_override
+
+    update_fields.extend(
+        [
+            "is_price_overridden",
+            "undiscounted_base_unit_price_amount",
+            "base_unit_price_amount",
+            "undiscounted_unit_price_gross_amount",
+            "undiscounted_unit_price_net_amount",
+        ]
+    )
 
 
 def add_gift_cards_to_order(
@@ -511,6 +556,7 @@ def change_order_line_quantity(
     channel: "Channel",
     manager: "PluginsManager",
     send_event=True,
+    update_fields=None,
 ):
     """Change the quantity of ordered items in a order line."""
     line = line_info.line
@@ -538,15 +584,17 @@ def change_order_line_quantity(
         line.undiscounted_total_price_net_amount = (
             undiscounted_total_price_net_amount.quantize(Decimal("0.001"))
         )
-        line.save(
-            update_fields=[
-                "quantity",
-                "total_price_net_amount",
-                "total_price_gross_amount",
-                "undiscounted_total_price_gross_amount",
-                "undiscounted_total_price_net_amount",
-            ]
-        )
+        fields = [
+            "quantity",
+            "total_price_net_amount",
+            "total_price_gross_amount",
+            "undiscounted_total_price_gross_amount",
+            "undiscounted_total_price_net_amount",
+        ]
+        if update_fields:
+            update_fields.extend(fields)
+        else:
+            line.save(update_fields=fields)
     else:
         delete_order_line(line_info, manager)
 
@@ -854,6 +902,7 @@ def update_discount_for_order_line(
     if reason is not None:
         order_line.unit_discount_reason = reason
         fields_to_update.append("unit_discount_reason")
+
     if current_value != value or current_value_type != value_type:
         undiscounted_base_unit_price = order_line.undiscounted_base_unit_price
         currency = undiscounted_base_unit_price.currency
@@ -868,6 +917,7 @@ def update_discount_for_order_line(
 
         order_line.unit_discount_type = value_type
         order_line.unit_discount_value = value
+        # TODO: should we save those values?
         order_line.total_price = order_line.unit_price * order_line.quantity
         order_line.undiscounted_unit_price = (
             order_line.unit_price + order_line.unit_discount
@@ -912,7 +962,7 @@ def _update_manual_order_line_discount_object(
     for discount in discounts:
         if discount.type == DiscountType.MANUAL and not discount_to_update:
             discount_to_update = discount
-        else:
+        elif discount.type != DiscountType.VOUCHER:
             discount_to_delete_ids.append(discount.pk)
 
     if discount_to_delete_ids:
@@ -929,6 +979,7 @@ def _update_manual_order_line_discount_object(
             amount_value=amount_value,
             currency=currency,
             reason=reason,
+            unique_type=DiscountType.MANUAL,
         )
     else:
         update_fields = []
@@ -1197,3 +1248,21 @@ def calculate_order_granted_refund_status(
 
     if with_save and current_granted_refund_status != granted_refund.status:
         granted_refund.save(update_fields=["status"])
+
+
+def log_address_if_validation_skipped_for_order(order: "Order", logger):
+    address = get_address_for_order_taxes(order)
+    if address and address.validation_skipped:
+        logger.warning(
+            "Fetching tax data for order with address validation skipped. "
+            "Address ID: %s",
+            address.id,
+        )
+
+
+def get_address_for_order_taxes(order: "Order"):
+    if order.collection_point_id:
+        address = order.collection_point.address  # type: ignore[union-attr]
+    else:
+        address = order.shipping_address or order.billing_address
+    return address